Cleared all the Datahandling warnings on my machine re #2818

+++ Code/Mantid/Framework/DataHandling/inc/MantidDataHandling/LoadPreNeXusMonitors.h
@@ -69,7 +69,7 @@ private:
   void exec();
 
   /// Number of monitors
-  size_t nMonitors;
+  int nMonitors;
 
   /// Set to true when instrument geometry was loaded.
   bool instrument_loaded_correctly;
